Refactored block element navigation to make less complex with fewer edge cases.
